Custom transitions same arguments
I do not know if this should be included, but I faced an issue when my custom transitions does not receive same arguments as DefaultTransition and I really needed those. So I came up with this solution. Please check this carefully, I will make fixed if needed or add more code to {$field}->transitionTo() as if used directly it wont have those arguments attached.