Recommend not to use hypens in the name
Hypens conflict with finding the transitionable states as they are used in the allowed transitions keys in the StateConfig class, they look something like 'fromState-toState', and this string is exploded at the hypen. If you have a hypen in the name, like "from-state-to-state" we confuse the system as the hyphen wont divide the class into valid states. (view StateConfig class, line 89).